renamed eth address component
